saveloc + replay fix
The replay system is currently buggy. If you finish the map, and then finish again with a worst time, your replay will be overwritten.

This fixes the replay save logic, and also saveloc by dont applying the singleZoneJump to players without an active timer.

Im using this on my server, and its working fine.